requirements.txt
# Core
torch>=2.0.0
transformers>=4.35.0
huggingface_hub>=0.20.0
accelerate>=0.25.0

# Server
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
Pillow>=10.0.0

# System monitoring (optional, for memory check)
psutil>=5.9.0

# Analysis (optional)
numpy>=1.24.0
pandas>=2.0.0
